+++ src/Attributes/Table.php
@@ -7,7 +7,7 @@ namespace SubstancePHP\SQL\Attributes;
 #[\Attribute(\Attribute::TARGET_CLASS)]
 class Table
 {
-    public function __construct(public string $name, public string $primaryKey)
+    public function __construct(public string $name)
     {
     }
 }